chore: update activity params
