+++ apps/app/src/query/index.ts
@@ -15,6 +15,9 @@
 
 export { useMutationWallet } from "./useMutationWallet";
 export { useMutationWalletSettings } from "./useMutationWalletSettings";
+export { useQueryActivities } from "./useQueryActivities";
+export { useQueryActivitiesCount } from "./useQueryActivitiesCount";
+export { useQueryConfiguration } from "./useQueryConfiguration";
 export { useSuspenseQueryAuthSession } from "./useSuspenseQueryAuthSession";
 export { useSuspenseQueryPortfolio } from "./useSuspenseQueryPortfolio";
 export { useSuspenseQueryTokenPrice } from "./useSuspenseQueryTokenPrice";
