+++ apps/app/src/query/useQueryActivitiesCount.ts
@@ -1,0 +1,74 @@
+// Copyright (C) 2023 Light, Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import { getActivitiesCount } from "@lightdotso/client";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
+import type { ActivityCountData } from "@/data";
+import type { ActivityListCountParams } from "@/params";
+import { queryKeys } from "@/queryKeys";
+import { useAuth } from "@/stores";
+
+export const useQueryActivitiesCount = (params: ActivityListCountParams) => {
+  // ---------------------------------------------------------------------------
+  // Stores
+  // ---------------------------------------------------------------------------
+
+  const { clientType } = useAuth();
+
+  // ---------------------------------------------------------------------------
+  // Query
+  // ---------------------------------------------------------------------------
+
+  const queryClient = useQueryClient();
+
+  const currentData: ActivityCountData | undefined = queryClient.getQueryData(
+    queryKeys.activity.listCount({ address: params.address }).queryKey,
+  );
+
+  const { data: activitiesCount } = useQuery<ActivityCountData | null>({
+    queryKey: queryKeys.activity.listCount({ address: params.address })
+      .queryKey,
+    queryFn: async () => {
+      if (!params.address) {
+        return null;
+      }
+
+      const res = await getActivitiesCount(
+        {
+          params: {
+            query: {
+              address: params.address,
+            },
+          },
+        },
+        clientType,
+      );
+
+      // Return if the response is 200
+      return res.match(
+        data => {
+          return data;
+        },
+        _ => {
+          return currentData ?? null;
+        },
+      );
+    },
+  });
+
+  return {
+    activitiesCount,
+  };
+};
